refactor: Voucher 관련 클래스 이름 변경 및 어노테이션 수정
