+++ src/main/java/com/programmers/springbootbasic/repository/voucher/MemoryVoucherRepository.java
@@ -1,12 +1,12 @@
-package com.programmers.springbootbasic.repository;
+package com.programmers.springbootbasic.repository.voucher;
 
 import com.programmers.springbootbasic.domain.voucher.Voucher;
-import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Repository;
 
 import java.util.*;
 
-@Component
-public class MemoryRepository implements Repository {
+@Repository
+public class MemoryVoucherRepository implements VoucherRepository {
     Map<UUID, Voucher> storage = new HashMap<>();
 
     public void save(Voucher voucher) {
