+++ src/main/java/com/programmers/springbootbasic/repository/voucher/VoucherRepository.java
@@ -1,10 +1,10 @@
-package com.programmers.springbootbasic.repository;
+package com.programmers.springbootbasic.repository.voucher;
 
 import com.programmers.springbootbasic.domain.voucher.Voucher;
 
 import java.util.List;
 
-public interface Repository {
+public interface VoucherRepository {
     void save(Voucher voucher);
 
     List<Voucher> findAll();
