+++ src/main/java/com/programmers/springbootbasic/service/VoucherService.java
@@ -2,30 +2,25 @@ package com.programmers.springbootbasic.service;
 
 import com.programmers.springbootbasic.domain.voucher.Voucher;
 import com.programmers.springbootbasic.domain.voucher.VoucherType;
-import com.programmers.springbootbasic.repository.Repository;
+import com.programmers.springbootbasic.repository.voucher.VoucherRepository;
 import com.programmers.springbootbasic.util.UuidProvider;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
+import lombok.AllArgsConstructor;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
 
-@Component
-public class Service {
-    private final Repository repository;
+@Service
+@AllArgsConstructor
+public class VoucherService {
+    private final VoucherRepository voucherRepository;
     private final UuidProvider uuidProvider;
-
-    @Autowired
-    public Service(Repository repository, UuidProvider uuidProvider) {
-        this.repository = repository;
-        this.uuidProvider = uuidProvider;
-    }
 
     public void create(VoucherType voucherType, Long amount) {
         Voucher voucher = voucherType.createVoucher(uuidProvider.generateUUID(), amount);
-        repository.save(voucher);
+        voucherRepository.save(voucher);
     }
 
     public List<Voucher> getAll() {
-        return repository.findAll();
+        return voucherRepository.findAll();
     }
 }
